requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
arrow==1.3.0
attrs==23.2.0
Authlib==1.3.0
autopep8==2.0.4
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
distro==1.9.0
eventkit==1.0.3
exceptiongroup==1.2.0
frozendict==2.4.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
ib-insync==0.9.86
idna==3.6
lxml==5.1.0
multitasking==0.0.11
nest-asyncio==1.5.8
numpy==1.26.3
openai==1.7.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
peewee==3.17.0
prompt-toolkit==3.0.43
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tda-api==1.6.0
tomli==2.0.1
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==12.0
wsproto==1.2.0
yfinance==0.2.35
